refactor: address review feedback on bounded search filter
- Cross-reference issue #4064 (JOIN spill-to-disk) as a reuse candidate
- Add KeySet::build_from(iterator) so the reader no longer drives the
  builder; make KeySetBuilder private
- Move per-statement warning helpers into a discoverable
  statement_warnings module alongside planner_warnings
- Reword the per-row-filter warning to "the table is being sequentially
  scanned for this query"
- Stop silencing the aggregate-scan diagnostic in the seq-scan regress
  test; surface it in context
