+++ pg_search/tests/pg_regress/expected/issue_3051.out
@@ -10,7 +10,7 @@ CALL paradedb.create_bm25_test_table(
 );
 CREATE INDEX on mock_items USING bm25 (id, description, rating, category, metadata) WITH (key_field='id', json_fields = '{"metadata": {"fast": true, "tokenizer": {"type": "raw", "lowercase": true}}}');
 SELECT id, description @@@ 'shoes' FROM mock_items ORDER BY id;
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id | ?column? 
 ----+----------
   1 | f
