+++ pg_search/tests/pg_regress/expected/or_exists_join_bug.out
@@ -140,7 +140,7 @@ AND EXISTS(
     )
 )
 ORDER BY u.id;
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id | name  
 ----+-------
   1 | Alice
@@ -168,7 +168,7 @@ AND EXISTS(
     )
 )
 ORDER BY u.id;
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id | name  
 ----+-------
   1 | Alice
@@ -195,7 +195,7 @@ AND EXISTS(
     )
 )
 ORDER BY u.id;
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id |  name   
 ----+---------
   3 | Charlie
@@ -249,7 +249,7 @@ AND EXISTS(
     )
 )
 ORDER BY u.id;
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  id | name  
 ----+-------
   1 | Alice
