+++ pg_search/tests/pg_regress/expected/snippet_json_02_advanced.out
@@ -75,7 +75,7 @@ LEFT JOIN reviews r ON r.book_id = bs.book_id
 WHERE r.id @@@ paradedb.parse('metadata.review:test') AND r.id @@@ paradedb.parse('metadata.review:snippet')
 ORDER BY bs.book_id, r.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b, r)
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |                            author_snippet                            | author_positions |                                   book_content_snippet                                    |  book_content_positions   | book_score | author_score |                                   review_text                                   |                                               review_snippet                                                |         review_positions         | review_score 
 ---------+----------------------------------------------------------------------+------------------+-------------------------------------------------------------------------------------------+---------------------------+------------+--------------+---------------------------------------------------------------------------------+-------------------------------------------------------------------------------------------------------------+----------------------------------+--------------
        1 |                                                                      |                  | This is a <b>test</b> <b>test</b> of the snippet function with multiple <b>test</b> words | {{10,14},{15,19},{58,62}} | 0.21010332 |            0 | This is a test review of the snippet function with multiple test words          | This is a <b>test</b> review of the <b>snippet</b> function with multiple <b>test</b> words                 | {{10,14},{29,36},{60,64}}        |   0.83736646
@@ -106,7 +106,7 @@ WHERE b.id @@@ paradedb.parse('metadata.content:test')
     OR r.id @@@ paradedb.parse('metadata.review:snippet')
 ORDER BY b.id, r.id;
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b, r)
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
  book_id |   author_name   |                                                                       book_snippet                                                                       |         book_positions          |                            author_snippet                            | author_positions |                                               review_snippet                                                |         review_positions         | book_score | author_score | review_score 
 ---------+-----------------+----------------------------------------------------------------------------------------------------------------------------------------------------------+---------------------------------+----------------------------------------------------------------------+------------------+-------------------------------------------------------------------------------------------------------------+----------------------------------+------------+--------------+--------------
        1 | Stephen King    | This is a <b>test</b> <b>test</b> of the snippet function with multiple <b>test</b> words                                                                | {{10,14},{15,19},{58,62}}       |                                                                      |                  | This is a <b>test</b> review of the <b>snippet</b> function with multiple <b>test</b> words                 | {{10,14},{29,36},{60,64}}        | 0.21010332 |            0 |   0.83736646
@@ -172,7 +172,7 @@ WHERE (a.id @@@ paradedb.parse('metadata.age:55'))
 ORDER BY a.id, b.id, r.id;
 WARNING:  JoinScan not used: LIMIT is required for top-level queries (tables: a, b)
 WARNING:  JoinScan not used: only INNER, ANTI, and SEMI JOIN are currently supported (types: LEFT, RIGHT) (tables: a, b, r)
-WARNING:  the BM25 index is not being used for this query; it is applied as a per-row filter, so query performance may be slow
+WARNING:  the table is being sequentially scanned for this query, so performance may be slow
 ERROR:  Unsupported query shape. Please report at https://github.com/paradedb/paradedb/issues/new/choose
 \i common/snippet_json_advanced_cleanup.sql
 DROP TABLE IF EXISTS authors;
